analysis/classification/requirements.txt
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
Cython==0.29.28
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.18
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==2.33.0
google-cloud-bigquery-storage==2.11.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.44.0
grpcio-status==1.44.0
hydralit-components==1.0.9
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
libcst==0.4.1
lxml==4.8.0
MarkupSafe==2.0.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg_resources==0.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.3.0
pydeck==0.6.2
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
streamlit==1.5.1
streamlit-aggrid==0.2.2.post2
streamlit-player==0.1.5
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==4.3.3
typing-inspect==0.7.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
youtube-transcript-api==0.4.3
zipp==3.6.0
pydata_google_auth==1.3.0